Create train_corrector.py
Training the corrector requires call to the command line tools provided by fairseq. The definition of the above functions used is further provided in the utility file. These functions require the important parameters which are appended to the command written for calling the inbuilt fairseq tools. The training process requires good code ( correct python code ) and respective bad code.